+++ app/src/main/java/ru/dubr/traineetestandroid/data/network/dto/PublicInterestStats.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class PublicInterestStats(
     @SerializedName("alexa_rank")
-    val alexaRank: Int,
+    val alexaRank: Double,
     @SerializedName("bing_matches")
     val bingMatches: Any
 )
